Function to add deployment labels

+++ src/OffalBot.Domain/ILabelMaker.cs
@@ -4,6 +4,6 @@ namespace OffalBot.Domain
 {
     public interface ILabelMaker
     {
-        Task CreateIfMissing(int repositoryId, string labelName, string labelColour);
+        Task CreateIfMissing(long repositoryId, string labelName, string labelColour);
     }
 }
