+++ src/OffalBot.Domain/ReviewRequest.cs
@@ -2,7 +2,7 @@ namespace OffalBot.Domain
 {
     public class ReviewRequest
     {
-        public int RepositoryId { get; set; }
+        public long RepositoryId { get; set; }
         public string PullRequestComment { get; set; }
         public string ReviewState { get; set; }
         public int PullRequestNumber { get; set; }
